Add end-to-end UMAP output-correctness tests; fix existing test types
umap-output.test.ts (new — 9 tests):
  Verifies the CPU pipeline produces a meaningful embedding, not just the
  right types/shapes. Uses two deterministic, well-separated clusters in
  6-D space with brute-force kNN (no WASM needed).

  Topology tests:
  - All embedding values finite (no NaN/Inf)
  - Embedding is not collapsed (x-std > 0.5)
  - kNN graph precondition: every neighbour is in the same cluster
  - Every point is closer to its own cluster centroid than the other's
  - Mean intra-cluster distance < ½ × mean inter-cluster distance
  - Nearest neighbour in 2-D embedding always belongs to the same cluster

  Bug-regression tests:
  - Bug 4: edges skip epoch 0 (epochOfNextSample starts at epochsPerSample)
  - Bug 4: edges fire correctly at epoch 1
  - Bug 6: LM-fitted a,b produce a curve within 0.1 of the smooth-step
           target at representative x values

cpu-sgd.test.ts:
  - FuzzyGraph rows/cols: Float32Array → Uint32Array (Bug 5 type fix)
  - 'attraction updates…' test: nEpochs 1→2 so the edge fires at epoch 1
    instead of epoch 0 (correct behaviour after Bug 4 fix); updated comment

cpu-sgd-transform.test.ts:
  - FuzzyGraph rows/cols: Float32Array → Uint32Array (Bug 5 type fix)

+++ src/__tests__/cpu-sgd-transform.test.ts
@@ -6,8 +6,8 @@ import { computeEps } from './test-helpers';
 /** Bipartite graph: new point 0 → training point 0, new point 1 → training point 1 */
 function makeBipartiteGraph(): FuzzyGraph {
   return {
-    rows: new Float32Array([0, 1]),   // new-point indices
-    cols: new Float32Array([0, 1]),   // training-point indices
+    rows: new Uint32Array([0, 1]),   // new-point indices
+    cols: new Uint32Array([0, 1]),   // training-point indices
     vals: new Float32Array([1.0, 1.0]),
     nVertices: 2,
   };
